restructure dflash code
