+++ src/speculators/models/dflash/core.py
@@ -1,5 +1,3 @@
-# ruff: noqa: ERA001
-
 from typing import TYPE_CHECKING, Any, ClassVar
 
 import torch
@@ -22,273 +20,17 @@ from typing_extensions import Unpack
 from speculators.model import SpeculatorModel
 from speculators.models.dflash import DFlashSpeculatorConfig
 from speculators.models.dflash.attention import create_anchor_block_mask_mod
+from speculators.models.dflash.metrics import compute_metrics
+from speculators.models.dflash.utils import (
+    _select_anchors,
+    build_kv_position_ids,
+    build_target_layer_ids,
+    gather_anchor_spans,
+)
 from speculators.utils.loading import load_model_layers
-
-
-@torch.no_grad()
-def compute_accuracy(
-    logits: torch.Tensor,  # shape: [1, total_seq_len - ttt_step, draft_vocab_size]
-    targets: torch.Tensor,  # shape: [1, total_seq_len - ttt_step, draft_vocab_size]
-    loss_mask: torch.Tensor | None,  # shape: [1, total_seq_len - ttt_step]
-    block_size: int = 1,
-):
-    # Note: logits, targets, and loss_mask are already aligned for the current ttt_step
-    # target_tokens = torch.argmax(targets, dim=-1)
-    target_tokens = targets
-    predicted_tokens = torch.argmax(logits, dim=-1)
-    correct = predicted_tokens == target_tokens
-
-    if block_size != 1:
-        accs = []
-        assert loss_mask is not None  # noqa: S101
-        for i in range(block_size):
-            pos_cor = torch.masked_select(
-                correct[:, i::block_size],
-                loss_mask.to(torch.bool)[:, i::block_size],
-            )
-            accs.append(pos_cor.float().sum() / (pos_cor.numel() + 1e-5))
-
-    if loss_mask is not None:
-        correct = torch.masked_select(correct, loss_mask.to(torch.bool))
-    correct_sum = correct.float().sum()
-    full_denom = correct.numel()
-
-    return correct_sum / (full_denom + 1e-5), accs
-
-
-def build_kv_position_ids(
-    base_position_ids: torch.Tensor,  # [B, total_seq_len]
-    anchor_positions: torch.Tensor,  # [B, n] or [n] (indices into base seq)
-    block_size: int,
-) -> torch.Tensor:
-    """
-    Construct position_ids for KV = [base | anchor_blocks].
-
-    Appended block for anchor a gets positions:
-        base_position_ids[..., a] + [0..block_size-1]
-    """
-    B, T = base_position_ids.shape  # noqa: N806
-    device = base_position_ids.device
-
-    # Normalize anchor_positions to [B, n]
-    if anchor_positions.ndim == 1:  # noqa: PLR2004
-        anchor_positions = (
-            anchor_positions.to(device=device, dtype=torch.long)
-            .unsqueeze(0)
-            .expand(B, -1)
-        )
-    elif anchor_positions.ndim == 2:  # noqa: PLR2004
-        anchor_positions = anchor_positions.to(device=device, dtype=torch.long)
-        if anchor_positions.shape[0] != B:
-            raise ValueError(
-                f"anchor_positions batch {anchor_positions.shape[0]} != {B}"
-            )
-    else:
-        raise ValueError(
-            f"anchor_positions must be [n] or [B, n], got {anchor_positions.shape}"
-        )
-
-    n = anchor_positions.shape[1]
-
-    # Position id at each anchor: [B, n]
-    anchor_pos_ids = torch.gather(
-        base_position_ids.to(torch.long), dim=1, index=anchor_positions
-    )
-
-    # Offsets within each block: [1, 1, block_size]
-    offsets = torch.arange(block_size, device=device, dtype=torch.long).view(
-        1, 1, block_size
-    )
-
-    # [B, n, block_size] -> [B, n*block_size]
-    appended_pos_ids = (anchor_pos_ids.unsqueeze(-1) + offsets).reshape(
-        B, n * block_size
-    )
-
-    return torch.cat([base_position_ids.to(torch.long), appended_pos_ids], dim=1)
-
-
-def gather_anchor_spans(
-    input_ids: torch.Tensor,
-    anchor_positions: torch.Tensor,
-    block_size: int,
-) -> torch.Tensor:
-    """
-    input_ids: [T]
-    anchor_positions: [n] (positions into input_ids)
-    returns: [n*block_size] = input_ids[anchor_i + 0 .. anchor_i +
-        block_size-1] concatenated
-    """
-    input_ids = input_ids.view(-1)
-    anchor_positions = anchor_positions.to(dtype=torch.long).view(-1)
-
-    # [block_size]
-    offsets = torch.arange(block_size, device=input_ids.device, dtype=torch.long)
-    # [n, block_size]
-    idx = anchor_positions[:, None] + offsets[None, :]
-
-    if (idx < 0).any() or (idx >= input_ids.numel()).any():
-        raise ValueError(
-            "Some anchor_positions + offsets are out of range for input_ids."
-        )
-
-    return input_ids[idx.reshape(-1)]
-
-
-def loss_function(logits, target_ids, loss_mask, block_size=8, gamma=4.0):
-    """
-    logits:     [B, T, V]  (draft vocab)
-    target_ids: [B, T]     (int64, in [0..V-1] or -100 for ignore)
-    loss_mask:  [B, T]     (0/1)
-    """
-    B, T, V = logits.shape  # noqa: N806
-
-    # per-token CE (no reduction yet)
-    ce = torch.nn.functional.cross_entropy(
-        logits.reshape(B * T, V),
-        target_ids.reshape(B * T),
-        reduction="none",
-        ignore_index=-100,
-    ).view(B, T)  # [B,T]
-
-    # aligned t -> original p=t+1 ; k = p % b (0 means anchor)
-    idx = torch.arange(T, device=logits.device)
-    k = (idx + 1) % block_size
-    w = torch.exp(-((k - 1).clamp(min=0)).to(logits.dtype) / gamma)
-    w = (w * (k != 0).to(logits.dtype)).view(1, T)  # anchors weight 0
-
-    m = loss_mask.to(logits.dtype).view(B, T)
-
-    ce = ce * w * m
-
-    denom = (m * w).sum(dim=1) + 1e-5
-    return (ce.sum(dim=1) / denom).mean()
-
-
-@torch.no_grad()
-def compute_acceptance_rate(
-    draft_logits: torch.Tensor,  # shape: [1, total_seq_len, draft_vocab_size]
-    target_logits: torch.Tensor,  # shape: [1, total_seq_len, draft_vocab_size]
-    loss_mask: torch.Tensor | None,  # shape: [1, total_seq_len]
-    block_size: int = 1,
-):
-    """
-    Compute acceptance rate for each position in the block according to
-    EAGLE 3 criteria.
-
-    EAGLE 3 acceptance formula: acceptance_prob = min(1, p / p_draft)
-    where p is the target model's probability and p_draft is the draft
-    model's probability.
-
-    Args:
-        draft_logits: Logits from the draft model
-        target_logits: Logits from the target/verifier model
-        loss_mask: Mask indicating which positions to include in the calculation
-        block_size: Size of each block for position-wise calculation
-
-    Returns:
-        If block_size == 1: overall acceptance rate
-        Otherwise: (overall acceptance rate, list of per-position acceptance rates)
-    """
-    # Convert logits to probabilities
-    draft_probs = torch.nn.functional.softmax(draft_logits, dim=-1)
-    target_probs = torch.nn.functional.softmax(target_logits, dim=-1)
-
-    # Get the draft tokens (what the draft model predicted)
-    draft_tokens = torch.argmax(draft_logits, dim=-1)
-
-    # Gather the probabilities for the draft tokens from both distributions
-    # Shape: [1, total_seq_len]
-    draft_token_probs_from_draft = torch.gather(
-        draft_probs, dim=-1, index=draft_tokens.unsqueeze(-1)
-    ).squeeze(-1)
-
-    draft_token_probs_from_target = torch.gather(
-        target_probs, dim=-1, index=draft_tokens.unsqueeze(-1)
-    ).squeeze(-1)
-
-    # Compute acceptance probability: min(1, p_target / p_draft)
-    # Add epsilon to avoid division by zero
-    acceptance_prob = torch.clamp(
-        draft_token_probs_from_target / (draft_token_probs_from_draft + 1e-10), max=1.0
-    )
-    accepted = acceptance_prob
-
-    if block_size != 1:
-        acc_rates = []
-        assert loss_mask is not None  # noqa: S101
-        for i in range(block_size):
-            pos_accepted = torch.masked_select(
-                accepted[:, i::block_size], loss_mask.to(torch.bool)[:, i::block_size]
-            )
-            acc_rates.append(pos_accepted.float().mean())
-
-    if loss_mask is not None:
-        accepted = torch.masked_select(accepted, loss_mask.to(torch.bool))
-
-    overall_acceptance = accepted.float().mean()
-
-    if block_size == 1:
-        return overall_acceptance
-    else:
-        return overall_acceptance, acc_rates
-
-
-def compute_metrics(
-    logits: torch.Tensor,
-    targets: torch.Tensor,
-    loss_mask: torch.Tensor | None,
-    block_size: int = 1,
-) -> tuple[torch.Tensor, dict]:
-    """Compute metrics for a given draft.
-
-    Args:
-        logits: The logits for the current ttt_step.
-        targets: The targets for the current ttt_step.
-        loss_mask: The loss mask for the current ttt_step.
-        prev_correct: The previous correct predictions for the current ttt_step.
-        ttt_step: The current ttt_step.
-        ttt_step_loss_decay: The loss decay for the current ttt_step.
-
-    Effects:
-        Modifies prev_correct in place.
-
-    Returns:
-        Loss value and metrics dictionary.
-    """
-    s_loss = loss_function(logits, targets, loss_mask)
-
-    s_full_acc, per_position_acc = compute_accuracy(
-        logits, targets, loss_mask, block_size
-    )
-    # s_accept_rate, per_position_accept = compute_acceptance_rate(
-    #     logits, targets, loss_mask, block_size
-    # )
-    s_metrics: dict[str, Any] = {}
-    s_metrics["loss"] = s_loss.detach().clone()
-    s_metrics["full_acc"] = s_full_acc
-    # s_metrics[f"accept_rate"] = s_accept_rate
-    for pos in range(len(per_position_acc)):
-        s_metrics[f"position {pos} acc"] = per_position_acc[pos]
-        # s_metrics[f"position {pos} accept"]=per_position_accept[pos]
-    return s_loss, s_metrics
-
 
 if TYPE_CHECKING:
     from collections.abc import Callable
-
-
-def build_target_layer_ids(num_target_layers: int, num_draft_layers: int):
-    """Calculate which target model layers to use for DFlash draft model."""
-    if num_draft_layers == 1:
-        return [(num_target_layers // 2)]
-    start = 1
-    end = num_target_layers - 3
-    span = end - start
-    return [
-        int(round(start + (i * span) / (num_draft_layers - 1)))
-        for i in range(num_draft_layers)
-    ]
 
 
 # Local copy of rotate_half to avoid dependency on internal transformers functions
@@ -299,7 +41,14 @@ def _rotate_half(x):
     return torch.cat((-x2, x1), dim=-1)
 
 
-def apply_rotary_pos_emb(q, k, cos, sin, position_ids=None, unsqueeze_dim=1):  # noqa: ARG001
+def apply_rotary_pos_emb(
+    q,
+    k,
+    cos,
+    sin,
+    position_ids=None,  # noqa: ARG001
+    unsqueeze_dim=1,
+):
     """Apply rotary position embeddings (local implementation)."""
 
     cos = cos.unsqueeze(unsqueeze_dim)
@@ -399,7 +148,9 @@ class Qwen3DFlashAttention(nn.Module):
             self.config._attn_implementation is not None  # noqa: SLF001
             and self.config._attn_implementation != "eager"  # noqa: SLF001
         ):
-            attn_fn = ALL_ATTENTION_FUNCTIONS[self.config._attn_implementation]  # noqa: SLF001
+            attn_fn = ALL_ATTENTION_FUNCTIONS[
+                self.config._attn_implementation  # noqa: SLF001
+            ]
         attn_output, attn_weights = attn_fn(
             self,
             q,
@@ -467,43 +218,13 @@ class Qwen3DFlashDecoderLayer(GradientCheckpointingLayer):
         return residual + hidden_states  # type: ignore[operator,return-value]
 
 
-def _select_anchors(
-    loss_mask: torch.Tensor, n: int, block_size: int
-) -> tuple[torch.Tensor, torch.Tensor]:
-    if loss_mask.ndim != 2:  # noqa: PLR2004
-        raise ValueError(f"Expected [B, T], got {loss_mask.shape}")
-
-    B, T = loss_mask.shape  # noqa: N806
-    valid_mask = loss_mask.bool().clone()
-
-    if block_size > 0:
-        valid_mask[:, T - block_size :] = False
-
-    out = []
-    out_valid = []
-    for b in range(B):
-        valid_indices = torch.nonzero(valid_mask[b], as_tuple=False).squeeze(1)
-
-        anchors = torch.zeros(n, dtype=torch.long, device=loss_mask.device)
-        anchor_valid = torch.zeros(n, dtype=torch.bool, device=loss_mask.device)
-
-        k = min(n, valid_indices.numel())
-        if k > 0:
-            perm = torch.randperm(valid_indices.numel(), device=loss_mask.device)
-            anchors[:k] = valid_indices[perm[:k]]
-            anchor_valid[:k] = True
-
-        out.append(anchors)
-        out_valid.append(anchor_valid)
-
-    return torch.stack(out, dim=0), torch.stack(out_valid, dim=0)
-
-
 @SpeculatorModel.register("dflash")
 class DFlashDraftModel(SpeculatorModel):
-    config_class: ClassVar[type[DFlashSpeculatorConfig]] = DFlashSpeculatorConfig  # type: ignore[misc]
+    # type: ignore[misc]
+    config_class: ClassVar[type[DFlashSpeculatorConfig]] = DFlashSpeculatorConfig
     _no_split_modules = ["Qwen3DFlashDecoderLayer"]
-    _keys_to_ignore_on_load_missing: ClassVar[list[str]] = ["embed_tokens.weight"]  # type: ignore[misc]
+    # type: ignore[misc]
+    _keys_to_ignore_on_load_missing: ClassVar[list[str]] = ["embed_tokens.weight"]
 
     def __init__(
         self,
@@ -517,11 +238,16 @@ class DFlashDraftModel(SpeculatorModel):
             verifier_attachment_mode="train_only",
         )
         self.config = config
-        # Set attention implementation to simple_flex_attention to support
-        # BlockMask
-        if self.config.transformer_layer_config._attn_implementation is None:  # noqa: SLF001
+        # Set attention implementation to simple_flex_attention
+        # to support BlockMask
+        if (
+            self.config.transformer_layer_config._attn_implementation  # noqa: SLF001
+            is None
+        ):
             impl = "simple_flex_attention"
-            self.config.transformer_layer_config._attn_implementation = impl  # noqa: SLF001
+            self.config.transformer_layer_config._attn_implementation = (  # noqa: SLF001
+                impl
+            )
         self.register_buffer("t2d", t2d)  # shape: [verifier_vocab_size], bool
         self.register_buffer("d2t", d2t)  # shape: [draft_vocab_size], int offsets
         self.draft_vocab_size = config.draft_vocab_size
@@ -533,7 +259,8 @@ class DFlashDraftModel(SpeculatorModel):
         num_draft_layers = config.transformer_layer_config.num_hidden_layers
         self.layers = nn.ModuleList(
             [
-                Qwen3DFlashDecoderLayer(config.transformer_layer_config, layer_idx)  # type: ignore[arg-type]
+                # type: ignore[arg-type]
+                Qwen3DFlashDecoderLayer(config.transformer_layer_config, layer_idx)
                 for layer_idx in range(num_draft_layers)
             ]
         )
@@ -556,7 +283,8 @@ class DFlashDraftModel(SpeculatorModel):
             config.transformer_layer_config.hidden_size,
             eps=config.transformer_layer_config.rms_norm_eps,
         )
-        self.rotary_emb = Qwen3RotaryEmbedding(config.transformer_layer_config)  # type: ignore[arg-type]
+        # type: ignore[arg-type]
+        self.rotary_emb = Qwen3RotaryEmbedding(config.transformer_layer_config)
 
         self.fc = nn.Linear(
             len(self.target_layer_ids) * config.transformer_layer_config.hidden_size,
@@ -710,8 +438,6 @@ class DFlashDraftModel(SpeculatorModel):
     @torch.compile
     def forward(
         self,
-        # shape: [1, total_seq_len, 5 * hidden_size]
-        # These are the hidden states from the target model.
         hidden_states: torch.Tensor,
         input_ids: torch.Tensor,
         lengths: torch.Tensor | None = None,  # shape: [batch_size]
@@ -727,7 +453,6 @@ class DFlashDraftModel(SpeculatorModel):
         if lengths is None:
             lengths = torch.tensor([total_seq_len], dtype=torch.long, device=device)
         if position_ids is None:
-            # MEGAN FLAG CHECK THAT THIS SHOULD BE +1
             position_ids = 1 + torch.arange(
                 total_seq_len, dtype=torch.long, device=device
             ).unsqueeze(0)
@@ -790,7 +515,7 @@ class DFlashDraftModel(SpeculatorModel):
                 attention_mask=attention_mask,
                 position_ids=position_ids,
                 past_key_value=past_key_values,
-                use_cache=False,  # FLAG MEGAN
+                use_cache=False,
                 position_embeddings=position_embeddings,
                 **kwargs,
             )
@@ -803,13 +528,11 @@ class DFlashDraftModel(SpeculatorModel):
             # t2d is a boolean mask [verifier_vocab_size] - True where
             # verifier token exists in draft
             # cumsum gives us the draft index for each verifier token
-            draft_indices = torch.cumsum(self.t2d.long(), dim=0) - 1  # type: ignore[operator]
+            # type: ignore[operator]
+            draft_indices = torch.cumsum(self.t2d.long(), dim=0) - 1
             targets_draft = torch.where(
-                # mask: is this verifier token in draft vocab?
                 self.t2d[targets],  # type: ignore[index]
-                # yes: get draft index
                 draft_indices[targets],
-                # no: ignore in loss
                 torch.tensor(-100, dtype=torch.long, device=device),
             )
